+++ Backend/users/managers.py
@@ -1,0 +1,29 @@
+from django.contrib.auth.base_user import BaseUserManager
+from django.db.models import Manager
+
+class CustomUserManager(BaseUserManager):
+    def create_user(self, email, password, **extra_fields):
+        if not email:
+            raise ValueError("The Email must be set")
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
+        user.set_password(password)
+        user.save()
+        return user
+
+    def create_superuser(self, email, password, **extra_fields):
+        extra_fields.setdefault("is_staff", True)
+        extra_fields.setdefault("is_superuser", True)
+        extra_fields.setdefault("is_active", True)
+
+        if extra_fields.get("is_staff") is not True:
+            raise ValueError(("Superuser must have is_staff=True."))
+        if extra_fields.get("is_superuser") is not True:
+            raise ValueError(("Superuser must have is_superuser=True."))
+        return self.create_user(email, password, **extra_fields)
+
+class EventManager(Manager):
+    def get_queryset(self):
+        query = super().get_queryset()
+        query = query.filter(soft_delete=False)
+        return query
