+++ Backend/users/models.py
@@ -1,27 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.base_user import BaseUserManager
+from .managers import *
 
-class CustomUserManager(BaseUserManager):
-    def create_user(self, email, password, **extra_fields):
-        if not email:
-            raise ValueError("The Email must be set")
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password, **extra_fields):
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-        extra_fields.setdefault("is_active", True)
-
-        if extra_fields.get("is_staff") is not True:
-            raise ValueError(("Superuser must have is_staff=True."))
-        if extra_fields.get("is_superuser") is not True:
-            raise ValueError(("Superuser must have is_superuser=True."))
-        return self.create_user(email, password, **extra_fields)
 
 def user_directory_path(instance, filename):
     # file will be uploaded to MEDIA_ROOT user_photo/ user_<id>/<filename>
@@ -127,6 +108,10 @@ class Event(models.Model):
     image = models.ImageField(null=True, blank=True)
     description = models.CharField(max_length=500)
     status = models.CharField(max_length=9, choices=STATUS, default="Draft")
+
+    # Handle soft deletion
+    soft_delete = models.BooleanField(default=False)
+    objects = EventManager()
 
     def __str__(self):
         return self.title
